feat: global variables
