+++ src/pages/cipp/settings/tabOptions.json
@@ -26,5 +26,9 @@
   {
     "label": "Licenses",
     "path": "/cipp/settings/licenses"
+  },
+  {
+    "label": "Global Variables",
+    "path": "/cipp/settings/global-variables"
   }
 ]
